Refactor header navigation more
